chore: make pcc happy
